+++ CHANGES.md
@@ -9,9 +9,11 @@ Release Notes.
 * Support Windows plugin test.
 * Support Kafka reporter.
 * Add recover to goroutine to prevent unexpected panics.
-* Add mutex to fix some data race. 
+* Add mutex to fix some data race.
 
 #### Plugins
+
+* Support [gozero](https://github.com/zeromicro/go-zero) Framework
 
 #### Documentation